+++ src/hooks/index.ts
@@ -1,3 +1,5 @@
 export { profileHook } from "./profileHook";
 
+export type { ProfileHookOptions } from "./profileHook";
+
 export { createHookProfiler } from "./createHookProfiler";